requirements.txt
flask
flask-restful
flask-sqlalchemy
flask-login
flask-marshmallow
flask-migrate
marshmallow-sqlalchemy
requests
